Refactored the testing for SyncWishlistCommand.
Created a WishlistParser for the logic for parsing items on the list, to separate features, with testing for this as well.

Also created a TODO, with concrete things I want to change and ideas for features.

Duplicate items also dont get added to wishlists now, as well as filtering by userId.